moving Hebrew text field to a different folder and creating a Hebrew text field type
